Add login and registration forms
